MISC: Remove models.py and routes.py to __init__ (duhhhh)
