+++ server/background/background.py
@@ -6,7 +6,6 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.engine import URL
 
-from models.model import Base
 from models.plant import Plant, Genus, Type
 from models.system import System, Light
 from models.alert import PlantAlert, Todo
