+++ server/db.py
@@ -7,7 +7,7 @@ from sqlalchemy.orm import sessionmaker, scoped_session
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.engine import URL
 
-from models.model import Base
+from models import Base
 from models.plant import Plant, Genus, Type
 from models.system import System, Light
 from models.alert import PlantAlert, Todo
